Cosmetic changes to product controller and enviroment variable file env

+++ server/controllers/productController.js
@@ -54,6 +54,9 @@ export const updatedProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).send("No product with that id");
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -62,6 +65,6 @@ export const deleteProduct = async (req, res) => {
   } catch (error) {
     console.log("Error in deleting product", error.message);
 
-    res.status(404).json({ success: false, message: "Product not found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
